Fixed issue with sort *IMPORTANT*
I'm finally 100% satisfied with the way my sort works. I originally wanted to make sorting update jsonbin however I realised there is no need for it especially when my method of deletion is with the date in milliseconds which is unique. This means that it doesn't matter in what way the todoList is shown on the page(sorted). The jsonbin will only update upon adding elements / deleting existing ones. This means that on every page reload the items will be sorted by date in ascending order which is now the default state of selectors. The issue I fixed is adding the correct date-ms to the corresponding todo-container's dataset to always delete the right one in jsonbin regardless of current sort shown.