Fixing tests by changing back default sorting method
So apparantly changing the default sorting method causes the 'sorting by priority' test to fail, so I changed the selectors back to sort by priority by default. Then I wanted to change the way todoList is sorted by default on page load(commented lines 32-35). This however causes the 'can add item in the end of the list' test to fail because it doesn't work if I sort the list on page load. I decided to leave it in as a comment because it's a cool feature but unfortunately I can't implement it and pass the tests.